Remove current state management approach
