--- lib/backend/state_models/carousel_model.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CarouselModel extends ChangeNotifier {
-  int _index = 0;
-
-  int get index => _index;
-
-  set index(int index) {
-    this._index = index;
-    notifyListeners();
-  }
-}
